+++ src/App.tsx
@@ -1,9 +1,7 @@
-import { useEffect, useState } from "react";
 import FeatureContents from "./components/global/FeatureContents";
 import Footer from "./components/global/Footer";
 import Header from "./components/global/Header";
 import { LiveEvenets } from "./components/global/LiveEvenets";
-import Navbar from "./components/global/Navbar";
 import NewsLetter from "./components/global/NewsLetter";
 import Privileges from "./components/global/Privileges";
 import SectionHead from "./components/global/SectionHead";
@@ -12,43 +10,10 @@ import { Testimonials } from "./components/global/Testimonials";
 
 export default function App() {
  
-  // navbar on scroll effect  
-  const [navStyle, setNavStyle] = useState("mt-2");
-  const [prevScrollPos, setPrevScrollPos] = useState(0);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const currentScrollPos = window.scrollY;
-      const isScrolledDown = prevScrollPos < currentScrollPos;
-
-      setPrevScrollPos(currentScrollPos);
-
-      // Show the navbar if scrolling up or at the top
-
-      if (window.scrollY > 10) {
-        if (!isScrolledDown) {
-          setNavStyle("mt-0 px-0 bg-white rounded-none");
-        } else {
-          setNavStyle("display-none  hidden");
-        }
-      } else {
-        setNavStyle("bg-none mt-2");
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [prevScrollPos]);
-
+  
 
   return (
-    <div className="min-h-[200vh]  overflow-hidden">
-      <div className={["fixed z-10 w-full   px-2 rounded-lg duration-300 transition-all", navStyle].join(" ")}>
-        <Navbar />
-      </div>
+    <>
       <Header title="Welcome to Kouralink" src="/src/assets/bg.png" />
       <div className="">
         <Privileges />
@@ -75,6 +40,6 @@ export default function App() {
       <SiteAnnouncements />
       <NewsLetter />
       <Footer />
-    </div>
+    </>
   );
 }
